Update test masters for first auto line number 0
The tests passed before this commit if I changed the first auto line
number back to 1, so I don't think any other issues are hiding in this
mass change. I did review the diff for the ASCII output as well, but I'm
taking it on faith the tokenised output is correct.